Integrate Crop class back into Focus class
For less breaking changes
